🐛 fix : 홈 화면 및 알람 QA 모두 수정
